Implements new Java Date API for Timespans
Changes representation of timestamps in millis from long to Instant.
Offers more flexibility and security.

+++ src/de/spreclib/api/main/FirstCentrifugationDurationOption.java
@@ -28,7 +28,7 @@ public final class FirstCentrifugationDurationOption extends AbstractListOption 
     return this.firstCentrifugationDuration;
   }
 
-  boolean hasDuration(int durationMinutes) {
+  boolean hasDuration(long durationMinutes) {
     if (this.firstCentrifugationDuration.hasValue(durationMinutes)) {
       return true;
     } else {
